// packages/upbit/src/generated/exchange/model/listDeposits200Item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Upbit Exchange API
 * 업비트 거래소 API. 공식 개발자 센터 v1.6.1 기준 스키마입니다.
 * OpenAPI spec version: 1.6.1
 */
import type { ListDeposits200ItemState } from './listDeposits200ItemState';
import type { ListDeposits200ItemTransactionType } from './listDeposits200ItemTransactionType';

export type ListDeposits200Item = {
  /** 입금 종류 */
  type?: string;
  /** 입금의 유일식별자(UUID) */
  uuid?: string;
  /** 조회하고자 하는 통화 코드 */
  currency?: string;
  /**
   * 입금 네트워크 유형.
업비트에서 사용하는 블록체인 네트워크 구분자입니다.

[예시] ETH, TRX, SOL

   * @nullable
   */
  net_type?: string | null;
  /** 입금 트랜잭션 ID.
 */
  txid?: string;
  /** 입금 처리 상태
  - `PROCESSING`: 입금 진행 중 (디지털 자산만 해당)
  - `ACCEPTED`: 완료
  - `CANCELLED`: 취소됨
  - `REJECTED`: 거절됨
  - `TRAVEL_RULE_SUSPECTED`: 트래블룰 추가 인증 대기중 (디지털 자산만 해당)
  - `REFUNDING`: 반환 절차 중
  - `REFUNDED`: 반환됨 (디지털 자산만 해당)
 */
  state?: ListDeposits200ItemState;
  /** 입금 요청 시각 (KST)

[형식] yyyy-MM-ddTHH:mm:ss+09:00
 */
  created_at?: string;
  /**
   * 입금 완료 시각 (KST)

[형식] yyyy-MM-ddTHH:mm:ss+09:00

   * @nullable
   */
  done_at?: string | null;
  /** 입금하고자 하는 원화의 금액.
 */
  amount?: string;
  /** 입금 수수료 */
  fee?: string;
  /** 입금 유형
  - `default`: 일반 입금
  - `internal`: 바로 입금 (업비트 계정간 입금)
 */
  transaction_type?: ListDeposits200ItemTransactionType;
};