// packages/upbit/src/generated/exchange/model/listWithdrawals200Item.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Upbit Exchange API
 * 업비트 거래소 API. 공식 개발자 센터 v1.6.1 기준 스키마입니다.
 * OpenAPI spec version: 1.6.1
 */
import type { ListWithdrawals200ItemState } from './listWithdrawals200ItemState';
import type { ListWithdrawals200ItemTransactionType } from './listWithdrawals200ItemTransactionType';

export type ListWithdrawals200Item = {
  /** 입출금 종류 */
  type: string;
  /** 출금의 유일 식별자 */
  uuid: string;
  /** 조회하고자 하는 통화 코드 */
  currency: string;
  /** 출금 네트워크 유형.
업비트에서 사용하는 블록체인 네트워크 구분자입니다. 원화(KRW) 출금의 경우 null로 반환됩니다.

[예시] "ETH", "TRX", "SOL"
 */
  net_type?: string;
  /**
   * 트랜잭션 ID
   * @nullable
   */
  txid: string | null;
  /** 출금 처리 상태
  - `WAITING`: 대기 중
  - `PROCESSING`: 처리 중
  - `DONE`: 완료
  - `FAILED`: 실패
  - `CANCELLED`: 취소됨
  - `REJECTED`: 거절됨
 */
  state: ListWithdrawals200ItemState;
  /** 출금 생성 시간 */
  created_at: string;
  /**
   * 출금 완료 시간. 출금이 완료되지 않은 경우 null로 반환됩니다
   * @nullable
   */
  done_at?: string | null;
  /** 출금하고자 하는 자산의 수량
 */
  amount: string;
  /** 출금 수수료 */
  fee: string;
  /** 출금 유형.
사용 가능한 값은 다음과 같습니다.

* `default`: 일반출금
* `internal`: 바로출금
 */
  transaction_type: ListWithdrawals200ItemTransactionType;
  /** 출금 취소 가능 여부 */
  is_cancelable: boolean;
};